requirements.txt
# Python dependencies for cloud computing assignment
# Task 1: Data Analysis
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
numpy>=1.21.0
jupyter>=1.0.0

# Task 3: Serverless function with Azure SDK
azure-storage-blob>=12.14.0

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0

# Optional: Enhanced data processing
scikit-learn>=1.1.0
plotly>=5.0.0